RHOAIENG-20088: chore(tests/containers): check images size change
Adds a new test to our pytest set to check the given image size compared
to the expected value. In this test we are checking the uncompressed
image size by summing up all the layeres of the image. This image is
usually downloaded on the machine where this is being run already.

https://issues.redhat.com/browse/RHOAIENG-20088