RHOAIENG-20088: chore(GHA CI): check images size change
Enhances our existing ci scripts for the params.env and existing
runtimes definitions with the actual image size check - we have preset
expected size for each of the image and we check that this size isn't
changed too much.

With this implementation we are checking the compressed image size to
avoid the image download.

https://issues.redhat.com/browse/RHOAIENG-20088